Refactor: 폴더 구조 개선 및 컴포넌트 역할별 분리
- components 디렉토리 내 구조를 역할 기반(DominoCanvas, DominoHUD, Common 등)으로 재정비
- 중첩 컴포넌트를 각각의 폴더로 분리하고 index.js를 추가하여 Barrel Export 적용
- 공통 훅(hooks), 유틸(utils), 상수(constants) 파일을 기능별 디렉토리로 분리
- 전체 프로젝트 구조를 도메인 중심으로 리팩토링하여 유지보수성과 가독성 향상

#60
