+++ src/components/DominoCanvas/CameraControls/CameraControls.jsx
