+++ src/components/DominoCanvas/ObjectRenderer/ObjectRenderer.jsx
