+++ src/components/DominoCanvas/index.js
@@ -1,0 +1,5 @@
+export { default as Ground } from "@/components/DominoCanvas/Ground/Ground";
+export { default as CameraControls } from "@/components/DominoCanvas/CameraControls/CameraControls";
+export { default as CursorFollowerObject } from "@/components/DominoCanvas/CursorFollowerObject/CursorFollowerObject";
+export { default as ObjectRenderer } from "@/components/DominoCanvas/ObjectRenderer/ObjectRenderer";
+export { default as DominoCanvas } from "@/components/DominoCanvas/DominoCanvas";
