+++ src/main.jsx
@@ -3,7 +3,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 
-import GlobalPortal from "@/GlobalPortal";
+import GlobalPortal from "@/components/Common/GlobalPortal";
 import routes from "@/routers/routes";
 
 const root = document.getElementById("root");
